move _dynamic_dict to DatasetBase
